requirements.txt
aiohttp
datasets
matplotlib
numpy
pandas
pillow
pyarrow
requests
streamlit
tqdm
transformers
torch
torchvision
torchaudio
